% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{adc3}
\alias{adc3}
\title{Areal depeletion curve using a 3 parameter model
\code{a*x^b+(1.0-a)*x^c}}
\usage{
adc3(a, b, c)
}
\arguments{
\item{a}{a parameter (0<a<1)}

\item{b}{b parameter (b>=0)}

\item{c}{c parameter (c>=0)}
}
\value{
11 element vector representing the ADC
}
\description{
Areal depeletion curve using a 3 parameter model
\code{a*x^b+(1.0-a)*x^c}
}
