% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_elev_curve}
\alias{area_elev_curve}
\title{Area Elevation Curve for the zones of SFLN2}
\format{
A data.table with
21 rows and 3 columns:
\describe{
\item{quantile}{percent area of the basin below the reference elevation}
\item{SFLN2-1}{reference elevations for zone 1}
\item{SFLN2-2}{reference elevations for zone 2}
}
}
\usage{
area_elev_curve
}
\description{
A dataset containing the percent area covered at a complete range of elevations
for a two zone configuration at SFLN2.
}
\keyword{datasets}
