% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{chanloss}
\alias{chanloss}
\title{Simple proportional chanloss}
\usage{
chanloss(flow, pars)
}
\arguments{
\item{pars}{parameters}

\item{flow_cfs}{streamflow vector}
}
\value{
Vector of flow scaled by the fixp parameter (if it exists)
}
\description{
Simple proportional chanloss
}
