% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forcingWGCM8}
\alias{forcingWGCM8}
\title{Forcing data for 2 zone model of WGCM8}
\format{
A list containing two data.frames (or data.table if you have that loaded).
each data frame has 43823 rows and 7 columns:
\describe{
\item{year,month,day,hour}{datetime columns}
\item{map_mm}{Mean areal precip in mm}
\item{mat_degc}{Mean areal temperature in degrees C}
\item{ptps}{Precent precip as snow}
}
}
\usage{
forcingWGCM8
}
\description{
A dataset containing 6 hour forcing data for the basin WGCM8 using 2 zones.
}
\keyword{datasets}
