% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{format_states}
\alias{format_states}
\title{Format state output from sac_snow_states}
\usage{
format_states(x)
}
\arguments{
\item{x}{output list from sac_snow_states}
}
\value{
a data.frame with formatted output
}
\description{
Format state output from sac_snow_states
}
