% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{lagk}
\alias{lagk}
\title{Lag-K Routing for any number of upstream points}
\usage{
lagk(dt_hours, uptribs, pars, sum_routes = TRUE)
}
\arguments{
\item{dt_hours}{timestep in hours}

\item{uptribs}{a matrix where each column contains flow data (in cfs) for an upstream point}

\item{pars}{parameters}
}
\value{
vector of routed flows
}
\description{
Lag-K Routing for any number of upstream points
}
\examples{
NULL
}
