% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pars}
\alias{pars}
\title{Parameters for a model of CHTO3}
\format{
A data.frame (or data.table if you have that loaded) with
43823 rows and 4 columns:
\describe{
\item{name}{The parameter name}
\item{zone}{the zone name, matching the names of the forcing list elements}
\item{value}{parameter value}
\item{type}{The parameter type (model), either 'sac', 'snow', or 'uh'}
}
}
\usage{
pars
}
\description{
A dataset containing parameters for a 2 zone SAC-SMA/SNOW17/UNITHG
model of the basin CHTO3 (6 hour timestep).
}
\keyword{datasets}
