% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{pet_hs}
\alias{pet_hs}
\title{Daily Potential Evapotranspiration using Hargreaves-Semani equations}
\usage{
pet_hs(lat, jday, tave, tmax, tmin)
}
\arguments{
\item{lat}{Latitude in decimal degrees}

\item{jday}{Julian day (Day of year since Jan 1)}

\item{tave}{Average daily temperature (C)}

\item{tmax}{Max daily temperature (C)}

\item{tmin}{Min daily temerature (C)}
}
\value{
Daily PET (vectorized over all inputs)
}
\description{
Daily Potential Evapotranspiration using Hargreaves-Semani equations
}
