% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfchydromodels.R
\docType{package}
\name{rfchydromodels}
\alias{rfchydromodels}
\title{rfchydromodels: A package running several models used operationally by the RFC}
\description{
The foo package provides three categories of important functions:
SAC-SMA/SNOW17 hydro models, UH channel routing, Lag-K Hydrograph routing
}
\section{SAC-SMA/SNOW17}{

Stuff
}

\section{UH}{

Stuff
}

\section{Lag-K routing}{

Stuff
}

