% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{sac_snow_uh}
\alias{sac_snow_uh}
\title{Execute SAC-SMA, SNOW17 and UH with given parameters}
\usage{
sac_snow_uh(dt_hours, forcing, pars, forcing_adjust = TRUE, climo = NULL)
}
\arguments{
\item{dt_hours}{timestep in hours}

\item{forcing}{data frame with columns for forcing inputs}

\item{pars}{sac parameters}

\item{forcing_adjust}{does the parameter set include forcing adjustments}

\item{climo}{pre computed climotology}
}
\value{
Vector of routed flow in cfs
}
\description{
Execute SAC-SMA, SNOW17 and UH with given parameters
}
\examples{
data(forcing)
data(pars)
dt_hours = 6
flow_cfs = sac_snow_uh(dt_hours,forcing, pars)
}
