% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{scale_uplimit}
\alias{scale_uplimit}
\title{Find a reasonable scale upper limit for optimization}
\usage{
scale_uplimit(shape, dt_hours)
}
\arguments{
\item{dt_hours}{}
}
\value{

}
\description{
Find a reasonable scale upper limit for optimization
}
