% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{sfc_pressure}
\alias{sfc_pressure}
\title{Conputes surface pressure in hPa from a given elevation}
\usage{
sfc_pressure(elev)
}
\arguments{
\item{elev}{surface elevation in meters}
}
\value{
Surface pressure in hPa
}
\description{
Conputes surface pressure in hPa from a given elevation
}
\examples{
sp <- sfc_pressure(0)
}
