% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tciSAKW1}
\alias{tciSAKW1}
\title{Total channel inflow (TCI) i.e runoff for a model of SAKW1}
\format{
A data.frame (or data.table if you have that loaded) with
517 rows and 4 columns:
\describe{
\item{name}{The parameter name}
\item{zone}{the zone name, matching the names of the forcing list elements}
\item{value}{parameter value}
\item{type}{The parameter type (model)}
}
}
\usage{
tciSAKW1
}
\description{
A dataset containing parameters for a 2 zone NWSRFS model
model of the basin SAKW1 (6 hour timestep).
}
\keyword{datasets}
