% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{uh}
\alias{uh}
\title{Two parameter unit hydrograph routing for one or more basin zones}
\usage{
uh(dt_hours, tci, pars)
}
\arguments{
\item{dt_hours}{timestep in hours}

\item{tci}{channel inflow matrix, one column per zone}

\item{pars}{parameters}
}
\value{
Vector of routed flow in cfs
}
\description{
Two parameter unit hydrograph routing for one or more basin zones
}
\examples{
# flow_cfs = uh(dt_hours, tci, pars)
}
