% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{uh2p}
\alias{uh2p}
\title{R port of the nwsrfs UH fortran code}
\usage{
uh2p(shape, scale, timestep, max_len = 1000)
}
\arguments{
\item{shape}{gamma shape parameter}

\item{scale}{gamma scale parameter}

\item{timestep}{timestep in hours}

\item{max_len}{max length of the uh}
}
\value{

}
\description{
Returns ordinates for a 2 parameter (shape,scale) gamma unit hydrograph. The ordinates are
based on the given timestep (in hours). To match the Fortran code, a max length is used.
A warning is issued if the UH does not terminate before the given max length.
}
\examples{
dt = 6
shape = 2
scale = 1
y = uh2p(2,1,6)
x = seq(dt,dt*length(y),by=dt)
plot(x,y,t='l')
}
