% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{uh2p_cfs_in}
\alias{uh2p_cfs_in}
\title{Create a 2 parameter gamma unit hydrograph with units cfs/in}
\usage{
uh2p_cfs_in(shape, scale, timestep, area)
}
\arguments{
\item{shape}{gamma shape parameter}

\item{scale}{gamma scale parameter}

\item{timestep}{timestep in hours}

\item{area}{basin area in square miles}
}
\value{
stuff
}
\description{
Create a 2 parameter gamma unit hydrograph with units cfs/in
}
\examples{
dt <- 6
shape <- 2
scale <- 1
y <- uh2p_cfs_in(2, 1, 6, 1000)
x <- seq(dt, dt * length(y), by = dt)
plot(x, y, t = "l")
}
