% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{uh2p_get_scale}
\alias{uh2p_get_scale}
\title{Get the scale parameter from a 2 parameter gamma unit hydrograph given
the shape parameter and time of concentration.}
\usage{
uh2p_get_scale(shape, toc, dt_hours)
}
\arguments{
\item{shape}{gamma shape parameter}

\item{toc}{time of concentration (hours)}

\item{dt_hours}{UH timestep (hours)}
}
\value{
scalar scale parameter
}
\description{
Get the scale parameter from a 2 parameter gamma unit hydrograph given
the shape parameter and time of concentration.
}
\examples{
uh2p_get_scale(2, 50, 1)
}
