% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{uh2p_seek}
\alias{uh2p_seek}
\title{Objective function for finding a scale parameter given shape and toc}
\usage{
uh2p_seek(scale, shape, dt_hours, toc)
}
\arguments{
\item{toc}{}
}
\value{

}
\description{
Objective function for finding a scale parameter given shape and toc
}
