% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upflowSAKW1}
\alias{upflowSAKW1}
\title{Upstream flow upflowSAKW1}
\format{
A data.frame (or data.table if you have that loaded) with
517 rows and 4 columns:
\describe{
\item{name}{The parameter name}
\item{zone}{the zone name, matching the names of the forcing list elements}
\item{value}{parameter value}
\item{type}{The parameter type (model)}
}
}
\usage{
upflowSAKW1
}
\description{
A dataset
}
\keyword{datasets}
