% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{chanloss}
\alias{chanloss}
\title{Seasonal chanloss}
\usage{
chanloss(flow, forcing, dt_hours, pars)
}
\arguments{
\item{flow}{streamflow vector}

\item{forcing}{forcing data}

\item{dt_hours}{timestep in hours}

\item{pars}{parameters}
}
\value{
Vector of flow modified by the chanloss pattern
}
\description{
Seasonal chanloss
}
