% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{sac_snow}
\alias{sac_snow}
\title{Execute SAC-SMA, SNOW17, return total channel inflow per zone}
\usage{
sac_snow(dt_hours, forcing, pars, forcing_adjust = TRUE)
}
\arguments{
\item{dt_hours}{timestep in hours}

\item{forcing}{data frame with with columns for forcing inputs}

\item{pars}{sac parameters}

\item{forcing_adjust}{does the parameter set include forcing adjustments}
}
\value{
Matrix (1 column per zone) of unrouted channel inflow
}
\description{
Execute SAC-SMA, SNOW17, return total channel inflow per zone
}
\examples{
data(forcing)
data(pars)
dt_hours = 6
tci = sac_snow(dt_hours, forcing, pars)
}
