% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{sac_snow_uh}
\alias{sac_snow_uh}
\title{Execute SAC-SMA, SNOW17 and UH with given parameters}
\usage{
sac_snow_uh(forcing, sac_pars, snow_pars, uh_pars, fa_pars = NULL)
}
\arguments{
\item{forcing}{data frame with with columns for forcing inputs}

\item{sac_pars}{sac parameters}

\item{snow_pars}{snow parameters}

\item{uh_pars}{uh parameters}

\item{fa_pars}{(optional) uh parameters}
}
\value{
Vector of routed flow in cfs
}
\description{
Execute SAC-SMA, SNOW17 and UH with given parameters
}
