% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{sac_snow_uh_lagk}
\alias{sac_snow_uh_lagk}
\title{Execute SAC-SMA, SNOW17, UH and LAG-K with given parameters}
\usage{
sac_snow_uh_lagk(dt_hours, forcing, uptribs, pars)
}
\arguments{
\item{dt_hours}{timestep in hours}

\item{forcing}{data frame with columns for forcing inputs}

\item{uptribs}{data frame with columns for upstream flow data}

\item{pars}{sac parameters}
}
\value{
Vector of routed flow in cfs
}
\description{
Execute SAC-SMA, SNOW17, UH and LAG-K with given parameters
}
\examples{
\dontrun{
data(forcing)
data(pars)
data(upflows)
dt_hours = 6
flow_cfs = sac_snow_uh_lagk(dt_hours, forcing, uptribs, pars)
}
}
