% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{uh}
\alias{uh}
\title{Two parameter unit hydrograph routing for one or more basin zones}
\usage{
uh(
  dt_hours,
  tci,
  pars,
  sum_zones = TRUE,
  start_of_timestep = TRUE,
  backfill = TRUE
)
}
\arguments{
\item{dt_hours}{timestep in hours}

\item{tci}{channel inflow matrix, one column per zone}

\item{pars}{parameters}

\item{sum_zones}{should routed flows from multiple zones be added and returned as a vector, or
kept separate and returned as a matrix}

\item{start_of_timestep}{should the output flow data be shifted by one timestep to account for
forcing data that uses beginning of timestep labeling}

\item{backfill}{when start_of_timestep is TRUE, should the first value be duplicated}
}
\value{
Vector of routed flow in cfs
}
\description{
Two parameter unit hydrograph routing for one or more basin zones
}
\examples{
data(forcingSAKW1)
data(parsSAKW1)
dt_hours <- 6
tci <- sac_snow(dt_hours, forcingSAKW1, parsSAKW1)
flow_cfs <- uh(dt_hours, tci, parsSAKW1)

data(tciSAKW1)
flow_cfs <- uh(dt_hours, tciSAKW1, parsSAKW1)
}
