% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{uh}
\alias{uh}
\title{Two parameter unit hydrograph routing for one or more basin zones}
\usage{
uh(dt_hours, tci, pars, sum_zones = TRUE)
}
\arguments{
\item{dt_hours}{timestep in hours}

\item{tci}{channel inflow matrix, one column per zone}

\item{pars}{parameters}

\item{sum_zones}{should routed flows from multiple zones be added and returned as a vector, or
kept separate and returned as a matrix}
}
\value{
Vector of routed flow in cfs
}
\description{
Two parameter unit hydrograph routing for one or more basin zones
}
\examples{
data(forcing)
data(pars)
dt_hours = 6
tci = sac_snow(dt_hours,forcing, pars)
flow_cfs = uh(dt_hours, tci, pars)
}
