% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac-snow-uh.R
\name{uh2p_seek2}
\alias{uh2p_seek2}
\title{Objective function for finding a scale parameter given shape and toc}
\usage{
uh2p_seek2(scale, shape, dt_hours, toc)
}
\arguments{
\item{scale}{blah}

\item{shape}{blah}

\item{dt_hours}{blah}

\item{toc}{blah}
}
\value{
function value for root finding
}
\description{
Objective function for finding a scale parameter given shape and toc
}
